requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cfgv==3.3.1
distlib==0.3.4
Django==4.0.4
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
filelock==3.6.0
identify==2.4.11
nodeenv==1.6.0
Pillow==9.1.0
pipenv==2022.1.8
platformdirs==2.5.1
pre-commit==2.17.0
pycodestyle==2.8.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
virtualenv==20.13.2
virtualenv-clone==0.5.7
